addressing feedback, refactoring tests
